requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
chromadb==0.6.3
cryptography==44.0.2
firebase-admin==6.7.0
Flask==3.1.0
flask_cors==4.0.0
google-genai==1.8.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
python-dotenv==1.0.0
PyJWT==2.10.1
pyparsing==3.2.3
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
